create app.js edit schema
