+++ prisma/schema.prisma
@@ -20,7 +20,7 @@ model User {
   email        String            @unique
   password     String
   mobile       String            @unique
-  birthDate    DateTime
+  birthDate    DateTime          @db.DateTime()
   gender       String
   occupation   String
   address      String?
